+++ src/compas_blender/drawing.py
